pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grcx"
version = "0.1.0"
description = "GRCX — the regulatory compliance monitoring agent that never sleeps"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Neil Lowden", email = "neil@grcx.dev"}]
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "watchdog>=4.0.0",
    "pypdf>=4.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "cryptography>=42.0.0",
    "flask>=3.0.0",
]

[project.scripts]
grcx = "grcx.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["grcx*"]
